# Speed up tests by running repetitions in parallel