--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -35,6 +35,10 @@ jobs:
         with:
           go-version: '1.20'
       - name: Test
+        env:
+          # Override test repetition parallelism, since the default runner seem to have a single core.
+          # This shaves ~30 seconds off the CI test duration. 
+          F3_TEST_REPETITION_PARALLELISM: 2
         run: make test
 
   generate: