Speed up tests by ~2X through running simulator test-case repetitions in parallel. The repetition parallelism is by default limited to the number of CPUs. Optionally, it is overridable by setting `F3_TEST_REPETITION_PARALLELISM` environment variable.

Refactor tests to reduce duplicate code and run tests in parallel wherever possible.

Note, the repetition parallelism limit is overridden to 2 for the CI runners. This reduces the test workflow execution time by ~10% compared to default, since the default runners run in a single core.

For comparison, the following lists the test time differences across three configurations:

1) This branch, with default parallelism limit:
```
$ time go test -count=1 ./...

?       github.com/filecoin-project/go-f3       [no test files]
?       github.com/filecoin-project/go-f3/adversary     [no test files]
?       github.com/filecoin-project/go-f3/cmd/f3sim     [no test files]
?       github.com/filecoin-project/go-f3/gen   [no test files]
ok      github.com/filecoin-project/go-f3/blssig        1.927s
ok      github.com/filecoin-project/go-f3/certs 2.331s
ok      github.com/filecoin-project/go-f3/gpbft 1.451s
ok      github.com/filecoin-project/go-f3/sim   2.738s
ok      github.com/filecoin-project/go-f3/test  94.702s

949.92s user 36.05s system 1036% cpu 1:35.15 total
```

2) This branch, with parallelism limit of 1:
```
$ time F3_TEST_REPETITION_PARALLELISM=1 go test -count=1 ./...

?       github.com/filecoin-project/go-f3       [no test files]
?       github.com/filecoin-project/go-f3/adversary     [no test files]
?       github.com/filecoin-project/go-f3/cmd/f3sim     [no test files]
?       github.com/filecoin-project/go-f3/gen   [no test files]
ok      github.com/filecoin-project/go-f3/blssig        1.192s
ok      github.com/filecoin-project/go-f3/certs 0.393s
ok      github.com/filecoin-project/go-f3/gpbft 0.819s
ok      github.com/filecoin-project/go-f3/sim   1.510s
ok      github.com/filecoin-project/go-f3/test  157.910s

548.50s user 4.93s system 349% cpu 2:38.33 total
```

3) The current head of `main` at `e5c4b15` :
```
$ time go test -count=1 ./...
?       github.com/filecoin-project/go-f3       [no test files]
?       github.com/filecoin-project/go-f3/adversary     [no test files]
?       github.com/filecoin-project/go-f3/cmd/f3sim     [no test files]
?       github.com/filecoin-project/go-f3/gen   [no test files]
ok      github.com/filecoin-project/go-f3/blssig        2.593s
ok      github.com/filecoin-project/go-f3/certs 3.048s
ok      github.com/filecoin-project/go-f3/gpbft 4.438s
ok      github.com/filecoin-project/go-f3/sim   3.536s
ok      github.com/filecoin-project/go-f3/test  164.805s

549.18s user 4.29s system 334% cpu 2:45.24 total
```